ShopNgo/requirements.txt
Flask
Werkzeug
flask-pymongo
flask-cors
bcrypt
python-dotenv
pymongo[srv]
python-jose[cryptography]
requests
websockets
pyotp
firebase-admin
flask-jwt-extended
protobuf
grpcio

PyJWT==2.3.0

